magit-diff-type: Guard rev-eq call to fix range edge case
As of 296f8dc7 (magit-diff-type: Fix case when the HEAD is used as
RANGE, 2018-09-13), we use `magit-rev-eq' to check whether RANGE
resolves to HEAD to avoid classifying `git diff HEAD'-equivalent
output as `committed'.

That fix introduced another edge case.  `magit-rev-eq' calls
`magit-rev-verify' on each argument, which returns the first line of
`git rev-parse --verify'.  In the case of a range A..B, the standard
output spans multiple lines, with first line being B's hash and the
second being "^A-hash".  This means that `(magit-rev-eq "A..B"
"HEAD")', where B resolves to "HEAD", will return true, and
`magit-diff-type' will incorrectly classify the diff as either
`unstaged' or `staged'.

Avoid this misclassification by confirming that RANGE isn't actually a
range before passing it to `magit-rev-eq'.  An alternative would be to
update `magit-rev-eq' so that it doesn't get confused by a range, but
`magit-diff-type' is the only caller that appears to pass it a range.