magit-req-eq: Return nil if A and/or B aren't commits
`magit-rev-eq' used to call `magit-rev-verify' on each argument, which
returns the first line of `git rev-parse --verify'.  In the case of a
range A..B, the standard output spans multiple lines, with first line
being B's hash and the second being "^A-hash".  It also return with
exit status 128, but `magit-rev-verify' ignores that.

`magit-rev-eq' should only return t if A and B refer to the same
commit and a range never refers to any single commit, so we should
return nil in that case.  Do that by using `magit-rev-verify-commit'
instead.